+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Added
 - Fixed map background buttons
 - Added ``enable_filter`` prop for ``k-select`` component to enable an input box which content will filter options.
 - MapBox Settings will now be stored within localStorage
+- Added ESLint and Prettier for linting and formatting.
 
 Fixed
 =====